requirements.txt
asgiref==3.8.1
boto3==1.35.44
botocore==1.35.44
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cryptography==43.0.3
defusedxml==0.8.0rc2
distlib==0.3.9
dj-database-url==2.2.0
Django==5.1.2
django-cors-headers==4.5.0
django-debug-toolbar==4.4.6
django-extensions==3.2.3
django-ses==4.2.0
django-storages==1.14.4
django-templated-mail==1.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
filelock==3.16.1
flake8==7.1.1
gunicorn==23.0.0
idna==3.10
Jinja2==3.1.4
jmespath==1.0.1
MarkupSafe==3.0.2
mccabe==0.7.0
nulltype==2.3.1
oauthlib==3.2.2
packaging==24.1
plaid-python==27.0.0
platformdirs==4.3.6
psycopg2-binary==2.9.10
pycodestyle==2.12.1
pycparser==2.22
pyflakes==3.2.0
PyJWT==2.9.0
pyOpenSSL==24.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
s3transfer==0.10.3
six==1.16.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.3
virtualenv==20.27.0
Werkzeug==0.16.1
